pyproject.toml
[project]
name = "nostro-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "mt5manager>=5.0.4874",
    "nostro-mt5-manager-v1>=0.0.7",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]
